+++ src/main/java/com/ecommerce/educative/repository/TokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TokenRepository extends JpaRepository<AuthenticationToken, Integer> {
+public interface TokenRepository extends JpaRepository<AuthenticationToken, Long> {
     AuthenticationToken findByUser(User user);
     AuthenticationToken findByToken(String token);
 }
